requirements.txt
# PartSelect Case Study - Complete Requirements
# Python 3.10+

# WEB SCRAPING
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# BACKEND (FastAPI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# AI/LLM
# Deepseek API (compatible with OpenAI client)
openai>=1.3.0

# RAG & VECTOR DATABASE
chromadb>=0.4.18
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.10

# DATA PROCESSING
pandas>=2.1.0
numpy>=1.26.0

# UTILITIES
python-dotenv>=1.0.0
tenacity>=8.2.3  # Retry logic
httpx>=0.25.0  # Async HTTP client